WIP moving filters into discussion
