+++ dotcom-rendering/src/components/Discussion.tsx
@@ -1,5 +1,5 @@
 import { css } from '@emotion/react';
-import { joinUrl } from '@guardian/libs';
+import { joinUrl, storage } from '@guardian/libs';
 import { palette, space } from '@guardian/source-foundations';
 import { Button, SvgPlus } from '@guardian/source-react-components';
 import { useEffect, useState } from 'react';
@@ -7,10 +7,16 @@ import { getCommentContext } from '../lib/getCommentContext';
 import { revealStyles } from '../lib/revealStyles';
 import { useDiscussion } from '../lib/useDiscussion';
 import { palette as themePalette } from '../palette';
-import type { SignedInUser } from '../types/discussion';
+import type {
+	FilterOptions,
+	OrderByType,
+	PageSizeType,
+	SignedInUser,
+} from '../types/discussion';
 import { Comments } from './Discussion/Comments';
 import { Hide } from './Hide';
 import { SignedInAs } from './SignedInAs';
+import { initFiltersFromLocalStorage } from '../lib/getCommentContext';
 
 export type Props = {
 	discussionApiUrl: string;
@@ -47,6 +53,12 @@ const positionRelative = css`
 	position: relative;
 `;
 
+const rememberFilters = ({ threads, pageSize, orderBy }: FilterOptions) => {
+	storage.local.set('gu.prefs.discussion.threading', threads);
+	storage.local.set('gu.prefs.discussion.pagesize', pageSize);
+	storage.local.set('gu.prefs.discussion.order', orderBy);
+};
+
 const commentIdFromUrl = () => {
 	const { hash } = window.location;
 	if (!hash.includes('comment')) return;
@@ -55,6 +67,30 @@ const commentIdFromUrl = () => {
 	if (!commentId) return;
 
 	return parseInt(commentId, 10);
+};
+
+const initialiseFilters = ({
+	pageSizeOverride,
+	orderByOverride,
+	permalinkBeingUsed,
+	isClosedForComments,
+}: {
+	pageSizeOverride?: PageSizeType;
+	orderByOverride?: OrderByType;
+	permalinkBeingUsed: boolean;
+	isClosedForComments: boolean;
+}) => {
+	// passing isClosedForComments feels like a smell here... is there a better way to do this?
+	const localFilters = initFiltersFromLocalStorage(isClosedForComments);
+	return {
+		// Override if prop given
+		pageSize: pageSizeOverride ?? localFilters.pageSize,
+		orderBy: orderByOverride ?? localFilters.orderBy,
+		threads:
+			localFilters.threads === 'collapsed' && permalinkBeingUsed
+				? 'expanded'
+				: localFilters.threads,
+	};
 };
 
 export const Discussion = ({
@@ -75,6 +111,31 @@ export const Discussion = ({
 	const [hashCommentId, setHashCommentId] = useState<number | undefined>(
 		commentIdFromUrl(),
 	);
+	const [filters, setFilters] = useState<FilterOptions>(
+		initialiseFilters({
+			commentPageSize,
+			commentOrderBy,
+			permalinkBeingUsed:
+				hashCommentId !== undefined && !Number.isNaN(hashCommentId),
+			isClosedForComments,
+		}),
+	);
+
+	// If these override props are updated we want to respect them
+	useEffect(() => {
+		const newFilters = {
+			...filters,
+			orderBy: commentOrderBy ?? filters.orderBy,
+			pageSize: commentPageSize ?? filters.pageSize,
+		};
+		setFilters(newFilters);
+	}, [commentPageSize, commentOrderBy]);
+
+	useEffect(() => {
+		rememberFilters(filters);
+		// Filters also show when the view is not expanded but we want to expand when they're changed
+		setIsExpanded(true);
+	}, [filters]);
 
 	const { commentCount, isClosedForComments } = useDiscussion(
 		joinUrl(discussionApiUrl, 'discussion', shortUrlId),
@@ -160,11 +221,9 @@ export const Discussion = ({
 					<Comments
 						user={user}
 						baseUrl={discussionApiUrl}
-						pageSizeOverride={commentPageSize}
 						isClosedForComments={
 							!!isClosedForComments || !enableDiscussionSwitch
 						}
-						orderByOverride={commentOrderBy}
 						shortUrl={shortUrlId}
 						additionalHeaders={{
 							'D2-X-UID': discussionD2Uid,
@@ -180,6 +239,8 @@ export const Discussion = ({
 						idApiUrl={idApiUrl}
 						page={commentPage}
 						setPage={setCommentPage}
+						filters={filters}
+						setFilters={setFilters}
 					/>
 					{!isExpanded && (
 						<div id="discussion-overlay" css={overlayStyles} />
